Relocate function check_if_settings_not_default()
